% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diopt.R
\name{diopt}
\alias{diopt}
\title{Query DIOPT database}
\usage{
diopt(genes, inTax, outTax)
}
\arguments{
\item{genes}{A vector of gene identifiers. Anything that DIOPT accepts}

\item{inTax}{taxid of the species that the input genes are coming from}

\item{outTax}{taxid of the species that you are seeking homology}
}
\description{
Query DIOPT database (\url{https://www.flyrnai.org/cgi-bin/DRSC_orthologs.pl}) for orthologues.
DIOPT database uses multiple tools to find gene orthologues. Sadly they don't have an
API so this function queries by visiting the site and filling up the form while obeying
their robots.txt. All queries will take a minimum of 10 seconds due to crawl delay.
}
