% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneHistory.R
\name{getGeneHistory}
\alias{getGeneHistory}
\title{Download gene history file}
\usage{
getGeneHistory(destfile = NULL, justRead = FALSE)
}
\arguments{
\item{destfile}{Path of the output file. If NULL a temp file will be used}

\item{justRead}{If TRUE and destfile exists, it reads the file instead of 
downloading the latest one from NCBI}
}
\value{
A data frame with latest gene history information
}
\description{
Downloads and reads the gene history file from NCBI website. This file is needed for
other functions
}
