% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneHistory.R
\name{getHomologene}
\alias{getHomologene}
\title{Get the latest homologene file}
\usage{
getHomologene(destfile = NULL, justRead = FALSE)
}
\arguments{
\item{destfile}{Path of the output file. If NULL a temp file will be used}

\item{justRead}{If TRUE and destfile exists, it reads the file instead of 
downloading the latest one from NCBI}
}
\value{
A data frame with homology groups, gene ids and gene symbols
}
\description{
This function downloads the latest homologene file from NCBI. Note that Homologene
has not been updated since 2014 so the output will be identical to \code{\link{homologeneData}}
included in this package. This function is here for futureproofing purposes.
}
