% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/homologene.R
\name{homologene}
\alias{homologene}
\title{Get homologues of given genes}
\usage{
homologene(genes, inTax, outTax, db = homologene::homologeneData)
}
\arguments{
\item{genes}{A vector of gene symbols or NCBI ids}

\item{inTax}{taxid of the species that the input genes are coming from}

\item{outTax}{taxid of the species that you are seeking homology}

\item{db}{Homologene database to use.}
}
\description{
Given a list of genes and a taxid, returns a data frame inlcuding the genes and their corresponding homologues
}
\examples{
homologene(c('Eno2','17441'), inTax = 10090, outTax = 9606)
}
