% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/homologene.R
\docType{data}
\name{homologeneData}
\alias{homologeneData}
\title{List of gene homologues used by homologene functions}
\format{An object of class \code{data.frame} with 114668 rows and 3 columns.}
\usage{
homologeneData
}
\description{
List of gene homologues used by homologene functions
}
\keyword{datasets}

