% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/homologene.R
\docType{data}
\name{homologeneVersion}
\alias{homologeneVersion}
\title{Version of homologene used}
\format{An object of class \code{integer} of length 1.}
\usage{
homologeneVersion
}
\description{
Version of homologene used
}
\keyword{datasets}
