% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/homologene.R
\name{human2mouse}
\alias{human2mouse}
\title{Human/mouse wraper for homologene}
\usage{
human2mouse(genes)
}
\arguments{
\item{genes}{A vector of gene symbols}
}
\description{
Human/mouse wraper for homologene
}

