% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/homologene.R
\name{mouse2human}
\alias{mouse2human}
\title{Mouse/human wraper for homologene}
\usage{
mouse2human(genes)
}
\arguments{
\item{genes}{A vector of gene symbols}
}
\description{
Mouse/human wraper for homologene
}

