% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/homologene.R
\docType{data}
\name{taxData}
\alias{taxData}
\title{Names and ids of included species}
\format{An object of class \code{data.frame} with 21 rows and 2 columns.}
\usage{
taxData
}
\description{
Names and ids of included species
}
\keyword{datasets}
