% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diopt.R
\name{diopt}
\alias{diopt}
\title{Query DIOPT database}
\usage{
diopt(genes, inTax, outTax)
}
\arguments{
\item{genes}{A vector of gene identifiers. Anything that DIOPT accepts}

\item{inTax}{taxid of the species that the input genes are coming from}

\item{outTax}{taxid of the species that you are seeking homology}
}
\description{
Query DIOPT database
}
