% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/homologene.R
\docType{data}
\name{homologeneData}
\alias{homologeneData}
\title{homologeneData}
\format{An object of class \code{data.frame} with 275237 rows and 4 columns.}
\usage{
homologeneData
}
\description{
List of gene homologues used by homologene functions
}
\keyword{datasets}
