% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/homologeneData2.R
\docType{data}
\name{homologeneData2}
\alias{homologeneData2}
\title{homologeneData2}
\format{An object of class \code{data.frame} with 112937 rows and 4 columns.}
\usage{
homologeneData2
}
\description{
A modified copy of the homologene database where the gene IDs and symbols are updated when necesary. 
Last update: Wed Feb 13 15:30:44 2019
}
\keyword{datasets}
