% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/homologeneData2.R
\docType{data}
\name{homologeneData2}
\alias{homologeneData2}
\title{homologeneData2}
\format{An object of class \code{data.frame} with 268846 rows and 4 columns.}
\usage{
homologeneData2
}
\description{
A modified copy of the homologene database. Homologene was updated at 2014 and many of its gene IDs and
symbols are out of date. Here the IDs and symbols are replaced with their most current version
Last update: Sun Sep  8 06:35:32 2019
}
\keyword{datasets}
