% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/homologene.R
\name{human2mouse}
\alias{human2mouse}
\title{Human/mouse wraper for homologene}
\usage{
human2mouse(genes, db = homologene::homologeneData)
}
\arguments{
\item{genes}{A vector of gene symbols or NCBI ids}

\item{db}{Homologene database to use.}
}
\description{
Human/mouse wraper for homologene
}
\examples{
human2mouse(c('ENO2','4340'))
}
