% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/homologene.R
\name{mouse2human}
\alias{mouse2human}
\title{Mouse/human wraper for homologene}
\usage{
mouse2human(genes, db = homologene::homologeneData)
}
\arguments{
\item{genes}{A vector of gene symbols or NCBI ids}

\item{db}{Homologene database to use.}
}
\description{
Mouse/human wraper for homologene
}
\examples{
mouse2human(c('Eno2','17441'))
}
