% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneHistory.R
\name{updateIDs}
\alias{updateIDs}
\title{Update gene IDs}
\usage{
updateIDs(ids, gene_history)
}
\arguments{
\item{ids}{Gene ids}

\item{gene_history}{Gene history information, probably returned by  \code{\link{getGeneHistory}}}
}
\value{
A character vector. New ids for genes that changed ids, or "-" for discontinued genes.
the input itself.
}
\description{
Given a list of gene ids and gene history information, traces changes in the 
gene's name to get the latest valid ID
}
\examples{
\dontrun{
gene_history = getGeneHistory()
updateIDs(c("4340964", "4349034", "4332470", "4334151", "4323831"),gene_history)
}

}
